// webapp/src/types/EntityReference.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type EntityReference = { id: bigint, item_id: bigint, entity_id: bigint, original_text: string, start_offset: bigint, end_offset: bigint, confidence: number | null, created_at: string, };